+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Note that since we don't clearly distinguish between a public and private interf
 ## [Unreleased]
 
 - Fix missing ``super.componentWillUnmount()`` calls (@simeonborko)
+- Add support for Glycam saccharide names
 
 ## [v3.9.0] - 2022-05-30